+++ src/__tests__/babelTiming.test.js
@@ -1,4 +1,5 @@
 const path = require('path');
+const fs = require('fs');
 const babelTiming = require('../index').babelTiming;
 const FIXTURES = '__fixtures__';
 
@@ -15,6 +16,8 @@ const expectedResultsEntry = {
   plugins: expect.arrayContaining([expectedPluginsEntry]),
 };
 
+const expectedResults = expect.arrayContaining([expectedResultsEntry]);
+
 function getFileList(results) {
   return results.map(entry => entry.name);
 }
@@ -22,8 +25,7 @@ function getFileList(results) {
 describe('babelTiming', () => {
   it('results have expected shape', async () => {
     const results = await babelTiming([path.join(FIXTURES, 'file-*.js*')]);
-    const resultsEntry = results[0];
-    expect(resultsEntry).toEqual(expectedResultsEntry);
+    expect(results).toEqual(expectedResults);
   });
 
   it('return empty array when no files are found', async () => {
@@ -56,6 +58,21 @@ describe('babelTiming', () => {
     it('returns results matching expected pattern', async () => {
       const results = await babelTiming([path.join(FIXTURES, 'file-*.js*')]);
       expect(results.length).toBe(3);
+    });
+  });
+
+  describe('"output" option', () => {
+    describe('is "JSON"', () => {
+      it('save result as JSON at the path specified by "outputPath" option', async () => {
+        await babelTiming([path.join(FIXTURES, 'file-*.js*')], {
+          output: 'json',
+          outputPath: './the-results.json',
+        });
+        const expectedFilePath = path.join(process.cwd(), 'the-results.json');
+        const actual = JSON.parse(fs.readFileSync(expectedFilePath));
+        fs.unlinkSync(expectedFilePath);
+        expect(actual).toEqual(expectedResults);
+      });
     });
   });